server/requirements.txt
fastapi
uvicorn
python-multipart
aspose-cad
